added token and some fixing

+++ src/redux/water/operations.js
@@ -3,24 +3,36 @@ import axios from "axios";
 
 export const addWater = createAsyncThunk(
   "water/addWater",
-  async ({ amount, date }, { rejectWithValue }) => {
+  async ({ amount, date }, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+    if (!token) {
+      return thunkAPI.rejectWithValue("No token found");
+    }
+    axios.defaults.headers.common.Authorization = `Bearer ${token}`;
     try {
       const response = await axios.post("/water", { amount, date });
       return response.data;
-    } catch (error) {
-      return rejectWithValue(error.message);
+    } catch (e) {
+      return thunkAPI.rejectWithValue(e.message);
     }
   }
 );
 
 export const fetchWater = createAsyncThunk(
   "water/fetchWater",
-  async (_, { rejectWithValue }) => {
+  async (_, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+    if (!token) {
+      return thunkAPI.rejectWithValue("No token found");
+    }
+    axios.defaults.headers.common.Authorization = `Bearer ${token}`;
     try {
       const response = await axios.get("/water");
       return response.data;
-    } catch (error) {
-      return rejectWithValue(error.message);
+    } catch (e) {
+      return thunkAPI.rejectWithValue(e.message);
     }
   }
 );
@@ -28,11 +40,17 @@ export const fetchWater = createAsyncThunk(
 export const updateWater = createAsyncThunk(
   "water/updateWater",
   async ({ waterId, ...water }, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+    if (!token) {
+      return thunkAPI.rejectWithValue("No token found");
+    }
+    axios.defaults.headers.common.Authorization = `Bearer ${token}`;
     try {
       const response = await axios.patch(`/water/${waterId}`, water);
       return response.data;
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error.message);
+    } catch (e) {
+      return thunkAPI.rejectWithValue(e.message);
     }
   }
 );
@@ -40,11 +58,17 @@ export const updateWater = createAsyncThunk(
 export const deleteWater = createAsyncThunk(
   "water/deleteWater",
   async (waterId, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+    if (!token) {
+      return thunkAPI.rejectWithValue("No token found");
+    }
+    axios.defaults.headers.common.Authorization = `Bearer ${token}`;
     try {
       const response = await axios.delete(`/water/${waterId}`);
       return response.data;
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error.message);
+    } catch (e) {
+      return thunkAPI.rejectWithValue(e.message);
     }
   }
 );
